Removed Maidump method, added vpk creation steps (#99)
* Removed Maidump method, added vpk creation steps

Maidump is not recommended for Vita dumping anymore, not even in specific deep circles of Vita dumping. This is because Maidump is unstable, and has lead to games being modified to somewhat work. Maidump is not well-supported, even on real Vitas and Vita TVs. The usage of Maidump can and will lead to your original game files being corrupted, forcing a re-installation of the game from the PlayStation Store, where the other methods will not result in corruption this way.

The addition of VPK creation is straightforward. It assumes that you've created a dump, rather than installed a VPK from other sources. As Vita3K wants to support users dumping their own cartridges and PSN games, this helps users into an easier method of installing games on Vita3K legitimately.